refactor: move strategy type file
